E2E test: skip deletion of test data dir in CI (#8006)
Don't delete the test data dir in CI. This will allow a test case
running serially to get all the bootstrap extensions in place and leave
them behind for the large batch of parallel cases that follow.

Also, conditionally skip clone because qa-example-content will be cloned
by the first case and doesn't need a reclone at the start of the
parallel cases run.

Also, update the current version of Air in product.json.

### QA Notes

@:web @:win

+++ test/e2e/infra/test-runner/test-setup.ts
@@ -110,7 +110,8 @@ function initializeTestEnvironment(rootPath = process.env.ROOT_PATH || 'ROOT_PAT
  * Cleans and prepares the test data directory.
  */
 function prepareTestDataDirectory() {
-	if (fs.existsSync(TEST_DATA_PATH)) {
+	// skipping deletion if running in CI because extensions setup case needs to be able to leave behind its extensions
+	if (!process.env.CI && fs.existsSync(TEST_DATA_PATH)) {
 		rimraf.sync(TEST_DATA_PATH);
 	}
 	mkdirp.sync(TEST_DATA_PATH);
